requirements.txt
chromadb
google-generativeai
langchain
langchain-community
langchain-google-genai
numpy
pandas
pyyaml
python-dotenv
scikit-learn
seaborn
streamlit
tiktoken
webvtt-py